feat(spec): implement spec meta data copy without array specification and wildcard
