+++ src/plugins/spec/CMakeLists.txt
@@ -2,6 +2,6 @@ include (LibAddMacros)
 
 add_plugin (
 	spec
-	SOURCES spec.h spec.c
+	SOURCES spec-new.h spec-new.c
 	LINK_ELEKTRA elektra-ease elektra-meta elektra-globbing
 	ADD_TEST COMPONENT libelektra${SO_VERSION})
