+++ src/plugins/spec/spec-new.h
